+++ src/MobileApp-src/src/main/java/com/sacram/proxy/Socks5Server.kt
@@ -251,7 +251,7 @@ class Socks5Server(
                 ?.toList().orEmpty()
         }
         val addrs = try {
-            future.get(dnsTimeoutMs, TimeUnit.MILLISECONDS)
+            future.get(dnsTimeoutMs.toLong(), TimeUnit.MILLISECONDS)
         } catch (e: TimeoutException) {
             throw IOException("DNS resolution timed out for $host on egress network $net", e)
         } catch (e: Exception) {
